Small cleanups in Scraper
