+++ kifi-backend/modules/scraper/test/com/keepit/scraper/fetcher/FakeHttpFetcherModule.scala
@@ -1,7 +1,6 @@
 package com.keepit.scraper.fetcher
 
 import com.google.inject.{ Provides, Singleton }
-import com.keepit.scraper.HttpFetchStatus
 
 case class FakeHttpFetcherModule(urlToResponse: Option[PartialFunction[String, HttpFetchStatus]] = None) extends HttpFetcherModule {
